fix: Split controller onto three different because of smell
